updated folder structure, added a landing page when admin, cleaner, custumer log in insted of using home page
