+++ src/components/layout/NavbarAdmin.js
@@ -14,10 +14,10 @@ const NavbarAdmin = () => {
           <Typography
             variant="h4"
             component={Link}
-            to="/"
+            to="/landingpage"
             sx={{ flexGrow: 1, textDecoration: "none", color: "white" }}
           >
-            Städa fint - Admin
+            StädaFint - Admin
           </Typography>
 
           <Button color="inherit" component={Link} to="/admincleaners">
