+++ src/components/layout/NavbarCleaner.js
@@ -14,10 +14,10 @@ const NavbarCleaner = () => {
           <Typography
             variant="h4"
             component={Link}
-            to="/"
+            to="/landingpage"
             sx={{ flexGrow: 1, textDecoration: "none", color: "white" }}
           >
-            Städa fint - Employee
+            StädaFint - Employee
           </Typography>
           <Button color="inherit" component={Link} to="/cleaner">
             My Account
