+++ src/components/layout/NavbarStart.js
@@ -14,10 +14,10 @@ const NavbarStart = () => {
           <Typography
             variant="h4"
             component={Link}
-            to="/"
+            to="/landingpage"
             sx={{ flexGrow: 1, textDecoration: "none", color: "white" }}
           >
-            Städa fint
+            StädaFint
           </Typography>
 
           <Button color="inherit" component={Link} to="/RegisterUser">
