+++ src/components/routes.js
@@ -1,20 +1,22 @@
 import { Routes, Route } from "react-router-dom";
-import CleanerView from "./views/CleanerView";
-import CustomerView from "./views/CustomerView";
+import CleanerView from "./views/CleanerViews/CleanerView";
+import CustomerView from "./views/CustomerViews/CustomerView";
 import RegisterUser from "./views/RegisterUser";
 import AdminView from "./views/AdminViews/AdminView";
 import AdminCleanerView from "./views/AdminViews/AdminCleanerView";
 import AdminCustomerView from "./views/AdminViews/AdminCustomerView";
-import RegisterBooking from "./views/RegisterBooking";
+import RegisterBooking from "./views/CustomerViews/RegisterBooking";
 import Login from "./views/Login";
 import Home from "./views/Home";
-import CustomerBookings from "./views/CustomerBookings";
-import CleanerViewBookings from "./views/CleanerViewBookings";
+import CustomerBookings from "./views/CustomerViews/CustomerBookings";
+import CleanerViewBookings from "./views/CleanerViews/CleanerViewBookings";
+import LandingPage from "./views/LandingPage";
 
 function AppRouter() {
   return (
     <Routes>
       <Route path="/" element={<Home />} />
+      <Route path="/landingpage" element={<LandingPage />} />
       <Route path="/cleaner" element={<CleanerView />} />
       <Route path="/cleanerbookings" element={<CleanerViewBookings />} />
       <Route path="/customer" element={<CustomerView />} />
